refactor(active-memory): keep runtime modules bounded
