+++ extensions/active-memory/transcript.ts
@@ -11,36 +11,24 @@ import {
   normalizeLowercaseStringOrEmpty,
   normalizeOptionalString,
 } from "openclaw/plugin-sdk/string-coerce-runtime";
-import { truncateUtf16Safe } from "openclaw/plugin-sdk/text-utility-runtime";
 import { clampInt } from "./config.js";
 import {
-  extractTextContent,
-  normalizeActiveSummary,
   readExplicitMemoryEvidence,
   readStructuredMemoryEvidenceFromContent,
   readStructuredMemoryFailure,
   readStructuredMemoryFailureFromContent,
-  truncateSummary,
 } from "./prompt.js";
+import { extractTextContent } from "./query.js";
 import {
   DEFAULT_ACTIVE_MEMORY_TOOLS_ALLOW,
   DEFAULT_PARTIAL_TRANSCRIPT_MAX_CHARS,
   DEFAULT_TRANSCRIPT_READ_MAX_BYTES,
   DEFAULT_TRANSCRIPT_READ_MAX_LINES,
   LANCEDB_ACTIVE_MEMORY_TOOLS_ALLOW,
-  TERMINAL_MEMORY_SEARCH_POLL_INTERVAL_MS,
-  TIMEOUT_PARTIAL_DATA_GRACE_MS,
-  type ActiveMemoryPartialTimeoutError,
   type ActiveMemorySearchDebug,
   type ActiveMemoryTranscriptSource,
-  type ActiveRecallResult,
-  type RecallSubagentResult,
-  type TerminalMemorySearchResult,
-  type TerminalMemorySearchWatch,
   type TranscriptReadLimits,
 } from "./types.js";
-
-let timeoutPartialDataGraceMs = TIMEOUT_PARTIAL_DATA_GRACE_MS;
 
 function isUnavailableMemorySearchDebug(debug?: ActiveMemorySearchDebug): boolean {
   return Boolean(debug?.error);
@@ -442,555 +430,16 @@ function hasUsableMemoryResultInSessionRecord(
   return Boolean(normalizedContent) && explicitEvidence !== false && structuredEvidence !== false;
 }
 
-async function readActiveMemoryTranscriptState(
-  source: ActiveMemoryTranscriptSource | string,
-  limits?: TranscriptReadLimits,
-  toolsAllow?: readonly string[],
-): Promise<{
-  searchDebug?: ActiveMemorySearchDebug;
-  hasUsableMemoryResult: boolean;
-  hasUnavailableMemorySearchResult: boolean;
-}> {
-  let searchDebug: ActiveMemorySearchDebug | undefined;
-  let hasUsableMemoryResult = false;
-  let hasUnavailableMemorySearchResult = false;
-  await streamActiveMemoryTranscriptRecords({
-    source: typeof source === "string" ? fileTranscriptSource(source) : source,
-    limits,
-    onRecord: (record) => {
-      const debug = extractActiveMemorySearchDebugFromSessionRecord(record);
-      if (debug) {
-        searchDebug = debug;
-      }
-      hasUnavailableMemorySearchResult ||= hasUnavailableMemoryResultInSessionRecord(
-        record,
-        toolsAllow,
-      );
-      hasUsableMemoryResult ||= hasUsableMemoryResultInSessionRecord(record, toolsAllow);
-    },
-  });
-  return { searchDebug, hasUsableMemoryResult, hasUnavailableMemorySearchResult };
-}
-
-async function readActiveMemorySearchDebug(
-  source: ActiveMemoryTranscriptSource | string,
-  limits?: TranscriptReadLimits,
-): Promise<ActiveMemorySearchDebug | undefined> {
-  return (await readActiveMemoryTranscriptState(source, limits)).searchDebug;
-}
-
-async function readMergedActiveMemoryTranscriptState(params: {
-  sources: readonly ActiveMemoryTranscriptSource[];
-  toolsAllow: readonly string[];
-}): Promise<{
-  searchDebug?: ActiveMemorySearchDebug;
-  hasUsableMemoryResult: boolean;
-  hasUnavailableMemorySearchResult: boolean;
-}> {
-  let searchDebug: ActiveMemorySearchDebug | undefined;
-  let hasUsableMemoryResult = false;
-  let hasUnavailableMemorySearchResult = false;
-  const seen = new Set<string>();
-  for (const source of params.sources) {
-    const key =
-      source.kind === "runtime"
-        ? `runtime:${source.target.agentId ?? ""}:${source.target.sessionId}:${source.target.sessionKey}:${source.target.storePath ?? ""}:${source.target.threadId ?? ""}`
-        : `file:${source.sessionFile}`;
-    if (seen.has(key)) {
-      continue;
-    }
-    seen.add(key);
-    const state = await readActiveMemoryTranscriptState(source, undefined, params.toolsAllow);
-    searchDebug = state.searchDebug ?? searchDebug;
-    hasUsableMemoryResult ||= state.hasUsableMemoryResult;
-    hasUnavailableMemorySearchResult ||= state.hasUnavailableMemorySearchResult;
-  }
-  return { searchDebug, hasUsableMemoryResult, hasUnavailableMemorySearchResult };
-}
-
-async function readTerminalMemorySearchResult(
-  source: ActiveMemoryTranscriptSource,
-  limits?: TranscriptReadLimits,
-  toolsAllow?: readonly string[],
-): Promise<TerminalMemorySearchResult | undefined> {
-  // memory_get consumes a path discovered by another tool; it is not an
-  // independent fallback that should delay terminal unavailability.
-  const recallPathNames = new Set(
-    toolsAllow
-      ?.map((toolName) => normalizeLowercaseStringOrEmpty(toolName))
-      .filter((toolName) => toolName && toolName !== "memory_get"),
-  );
-  if (recallPathNames.size === 0) {
-    return undefined;
-  }
-  const unavailablePathNames = new Set<string>();
-  let hasUsableMemoryResult = false;
-  let searchDebug: ActiveMemorySearchDebug | undefined;
-  await streamActiveMemoryTranscriptRecords({
-    source,
-    limits,
-    onRecord: (record) => {
-      hasUsableMemoryResult ||= hasUsableMemoryResultInSessionRecord(record, toolsAllow);
-      searchDebug = extractActiveMemorySearchDebugFromSessionRecord(record) ?? searchDebug;
-      const toolName = extractToolResultNameFromSessionRecord(record);
-      if (!toolName || !recallPathNames.has(toolName)) {
-        return false;
-      }
-      if (hasTerminalUnavailableMemoryResultInSessionRecord(record, toolsAllow ?? [])) {
-        unavailablePathNames.add(toolName);
-      } else {
-        unavailablePathNames.delete(toolName);
-      }
-      return false;
-    },
-  });
-  if (unavailablePathNames.size !== recallPathNames.size) {
-    return undefined;
-  }
-  return {
-    status: "unavailable",
-    hasUsableMemoryResult,
-    searchDebug,
-  };
-}
-
-async function readTerminalMemorySearchResultFromSources(
-  sources: readonly ActiveMemoryTranscriptSource[],
-  limits: TranscriptReadLimits | undefined,
-  toolsAllow: readonly string[],
-): Promise<TerminalMemorySearchResult | undefined> {
-  for (const source of sources) {
-    const result = await readTerminalMemorySearchResult(source, limits, toolsAllow);
-    if (result) {
-      return result;
-    }
-  }
-  return undefined;
-}
-
-function watchTerminalMemorySearchResult(params: {
-  getTranscriptSources: () => readonly ActiveMemoryTranscriptSource[];
-  abortSignal: AbortSignal;
-  toolsAllow: readonly string[];
-}): TerminalMemorySearchWatch {
-  let stopped = false;
-  let timeoutId: ReturnType<typeof setTimeout> | undefined;
-  let inFlight = false;
-  let resolveWatch: (result: TerminalMemorySearchResult) => void = () => {};
-  const stop = () => {
-    if (stopped) {
-      return;
-    }
-    stopped = true;
-    if (timeoutId) {
-      clearTimeout(timeoutId);
-      timeoutId = undefined;
-    }
-    params.abortSignal.removeEventListener("abort", onAbort);
-  };
-  const finish = (result: TerminalMemorySearchResult) => {
-    stop();
-    resolveWatch(result);
-  };
-  const schedule = () => {
-    if (stopped) {
-      return;
-    }
-    timeoutId = setTimeout(() => {
-      void tick();
-    }, TERMINAL_MEMORY_SEARCH_POLL_INTERVAL_MS);
-    timeoutId.unref?.();
-  };
-  const tick = async () => {
-    if (stopped || inFlight) {
-      return;
-    }
-    if (params.abortSignal.aborted) {
-      stop();
-      return;
-    }
-    inFlight = true;
-    try {
-      const result = await readTerminalMemorySearchResultFromSources(
-        params.getTranscriptSources(),
-        undefined,
-        params.toolsAllow,
-      );
-      if (result) {
-        finish(result);
-        return;
-      }
-    } catch {
-      // Transcript polling is opportunistic; normal timeout handling remains authoritative.
-    } finally {
-      inFlight = false;
-    }
-    schedule();
-  };
-  function onAbort() {
-    stop();
-  }
-  const promise = new Promise<TerminalMemorySearchResult>((resolve) => {
-    resolveWatch = resolve;
-    params.abortSignal.addEventListener("abort", onAbort, { once: true });
-    void tick();
-  });
-  return {
-    promise,
-    stop,
-  };
-}
-
-function normalizeSearchDebug(value: unknown): ActiveMemorySearchDebug | undefined {
-  const debug = asRecord(value);
-  if (!debug) {
-    return undefined;
-  }
-  const normalized: ActiveMemorySearchDebug = {
-    backend: normalizeOptionalString(debug.backend),
-    configuredMode: normalizeOptionalString(debug.configuredMode),
-    effectiveMode: normalizeOptionalString(debug.effectiveMode),
-    fallback: normalizeOptionalString(debug.fallback),
-    searchMs:
-      typeof debug.searchMs === "number" && Number.isFinite(debug.searchMs)
-        ? debug.searchMs
-        : undefined,
-    hits: typeof debug.hits === "number" && Number.isFinite(debug.hits) ? debug.hits : undefined,
-    warning: normalizeOptionalString(debug.warning) ?? normalizeOptionalString(debug.reason),
-    action: normalizeOptionalString(debug.action),
-    error: normalizeOptionalString(debug.error),
-  };
-  return normalized.backend ||
-    normalized.configuredMode ||
-    normalized.effectiveMode ||
-    normalized.fallback ||
-    typeof normalized.searchMs === "number" ||
-    typeof normalized.hits === "number" ||
-    normalized.warning ||
-    normalized.action ||
-    normalized.error
-    ? normalized
-    : undefined;
-}
-
-function readActiveMemorySearchDebugFromRunResult(
-  result: unknown,
-): ActiveMemorySearchDebug | undefined {
-  const record = asRecord(result);
-  const meta = asRecord(record?.meta);
-  return (
-    normalizeSearchDebug(meta?.activeMemorySearchDebug) ??
-    normalizeSearchDebug(meta?.memorySearchDebug) ??
-    normalizeSearchDebug(record?.activeMemorySearchDebug) ??
-    normalizeSearchDebug(record?.memorySearchDebug)
-  );
-}
-
-function readActiveMemorySessionFileFromRunResult(result: unknown): string | undefined {
-  const record = asRecord(result);
-  const meta = asRecord(record?.meta);
-  const agentMeta = asRecord(meta?.agentMeta);
-  return (
-    normalizeOptionalString(agentMeta?.sessionFile) ?? normalizeOptionalString(meta?.sessionFile)
-  );
-}
-
-function readMemoryToolResultEvidence(params: {
-  toolName: string;
-  result: unknown;
-  isError: boolean;
-  toolsAllow: readonly string[];
-}): {
-  hasUsableMemoryResult: boolean;
-  hasUnavailableMemorySearchResult: boolean;
-} {
-  const result = asRecord(params.result);
-  const rawContent = result?.content;
-  const textContent =
-    normalizeOptionalString(result?.detailedContent) ??
-    (typeof rawContent === "string" ? normalizeOptionalString(rawContent) : undefined);
-  const record = {
-    message: {
-      role: "toolResult",
-      toolName: params.toolName,
-      isError: params.isError,
-      content: Array.isArray(rawContent)
-        ? rawContent
-        : textContent
-          ? [{ type: "text", text: textContent }]
-          : [],
-      details: result?.details,
-    },
-  };
-  return {
-    hasUsableMemoryResult: hasUsableMemoryResultInSessionRecord(record, params.toolsAllow),
-    hasUnavailableMemorySearchResult: hasUnavailableMemoryResultInSessionRecord(
-      record,
-      params.toolsAllow,
-    ),
-  };
-}
-
-function extractAssistantTextFromSessionRecord(value: unknown): string {
-  const record = asRecord(value);
-  if (!record) {
-    return "";
-  }
-  const nestedMessage = asRecord(record.message);
-  const topLevelMessage = normalizeOptionalString(record.role) === "assistant" ? record : undefined;
-  const message = nestedMessage ?? topLevelMessage;
-  if (!message || normalizeOptionalString(message.role) !== "assistant") {
-    return "";
-  }
-  return extractTextContent(message.content).trim();
-}
-
-async function readPartialAssistantText(
-  source: ActiveMemoryTranscriptSource | string | undefined,
-  limits?: TranscriptReadLimits,
-): Promise<string | null> {
-  if (!source) {
-    return null;
-  }
-  const texts: string[] = [];
-  const resolvedLimits = resolveTranscriptReadLimits(limits);
-  let collectedChars = 0;
-  await streamActiveMemoryTranscriptRecords({
-    source: typeof source === "string" ? fileTranscriptSource(source) : source,
-    limits: resolvedLimits,
-    onRecord: (record) => {
-      const text = extractAssistantTextFromSessionRecord(record);
-      if (text) {
-        const separatorChars = texts.length > 0 ? 1 : 0;
-        const remaining = resolvedLimits.maxChars - collectedChars - separatorChars;
-        if (remaining <= 0) {
-          return true;
-        }
-        const nextText = truncateUtf16Safe(text, remaining);
-        if (!nextText) {
-          return true;
-        }
-        texts.push(nextText);
-        collectedChars += separatorChars + nextText.length;
-        // A surrogate backoff leaves spare code units; stop instead of skipping ahead.
-        return nextText.length < text.length || collectedChars >= resolvedLimits.maxChars;
-      }
-      return false;
-    },
-  });
-  // Accepted chunks and separators are charged before append, so the join is already bounded.
-  const joined = texts.join("\n").trim();
-  return joined || null;
-}
-
-async function readPartialAssistantTextFromSources(
-  sources: readonly ActiveMemoryTranscriptSource[],
-  limits?: TranscriptReadLimits,
-): Promise<string | null> {
-  for (const source of sources) {
-    const text = await readPartialAssistantText(source, limits);
-    if (text) {
-      return text;
-    }
-  }
-  return null;
-}
-
-function attachPartialTimeoutData(
-  error: unknown,
-  partialReply: string | null,
-  searchDebug: ActiveMemorySearchDebug | undefined,
-  hasUnavailableMemorySearchResult: boolean,
-): void {
-  if (!error || typeof error !== "object") {
-    return;
-  }
-  const target = error as ActiveMemoryPartialTimeoutError;
-  if (partialReply) {
-    target.activeMemoryPartialReply = partialReply;
-  }
-  if (searchDebug) {
-    target.activeMemorySearchDebug = searchDebug;
-  }
-  if (hasUnavailableMemorySearchResult) {
-    target.activeMemoryUnavailableMemorySearch = true;
-  }
-}
-
-function readPartialTimeoutData(error: unknown): {
-  rawReply?: string;
-  searchDebug?: ActiveMemorySearchDebug;
-  hasUnavailableMemorySearchResult?: boolean;
-} {
-  if (!error || typeof error !== "object") {
-    return {};
-  }
-  const source = error as ActiveMemoryPartialTimeoutError;
-  return {
-    rawReply: normalizeOptionalString(source.activeMemoryPartialReply),
-    searchDebug: source.activeMemorySearchDebug,
-    hasUnavailableMemorySearchResult: source.activeMemoryUnavailableMemorySearch,
-  };
-}
-
-async function waitForSubagentPartialTimeoutData(
-  subagentPromise: Promise<RecallSubagentResult> | undefined,
-): Promise<{
-  rawReply?: string;
-  searchDebug?: ActiveMemorySearchDebug;
-  hasUnavailableMemorySearchResult?: boolean;
-  settled: boolean;
-}> {
-  if (!subagentPromise) {
-    return { settled: true };
-  }
-  let timeoutId: ReturnType<typeof setTimeout> | undefined;
-  const timeoutPromise = new Promise<{ settled: false }>((resolve) => {
-    timeoutId = setTimeout(() => resolve({ settled: false }), timeoutPartialDataGraceMs);
-    timeoutId.unref?.();
-  });
-  try {
-    return await Promise.race([
-      subagentPromise.then(
-        () => ({ settled: true as const }),
-        (error: unknown) => ({ ...readPartialTimeoutData(error), settled: true as const }),
-      ),
-      timeoutPromise,
-    ]);
-  } finally {
-    if (timeoutId) {
-      clearTimeout(timeoutId);
-    }
-  }
-}
-
-async function buildTimeoutRecallResult(params: {
-  elapsedMs: number;
-  maxSummaryChars: number;
-  transcriptSources: readonly ActiveMemoryTranscriptSource[];
-  rawReply?: string;
-  searchDebug?: ActiveMemorySearchDebug;
-  hasUnavailableMemorySearchResult?: boolean;
-  subagentPromise?: Promise<RecallSubagentResult>;
-  toolsAllow: readonly string[];
-}): Promise<ActiveRecallResult> {
-  const subagentPartialData = params.rawReply
-    ? { settled: true as const }
-    : await waitForSubagentPartialTimeoutData(params.subagentPromise);
-  const rawReply =
-    params.rawReply ??
-    subagentPartialData.rawReply ??
-    (await readPartialAssistantTextFromSources(params.transcriptSources));
-  const summary = truncateSummary(
-    normalizeActiveSummary(rawReply ?? "") ?? "",
-    params.maxSummaryChars,
-  );
-  const transcriptState =
-    params.transcriptSources.length > 0
-      ? await readMergedActiveMemoryTranscriptState({
-          sources: params.transcriptSources,
-          toolsAllow: params.toolsAllow,
-        })
-      : undefined;
-  const searchDebug =
-    params.searchDebug ?? subagentPartialData.searchDebug ?? transcriptState?.searchDebug;
-  if (
-    summary.length === 0 ||
-    isUnavailableMemorySearchDebug(searchDebug) ||
-    !subagentPartialData.settled ||
-    params.hasUnavailableMemorySearchResult ||
-    subagentPartialData.hasUnavailableMemorySearchResult ||
-    transcriptState?.hasUnavailableMemorySearchResult
-  ) {
-    return {
-      status: "timeout",
-      elapsedMs: params.elapsedMs,
-      summary: null,
-      searchDebug,
-    };
-  }
-  return {
-    status: "timeout_partial",
-    elapsedMs: params.elapsedMs,
-    summary,
-    searchDebug,
-  };
-}
-
-function buildSubagentRecallResult(params: {
-  subagentResult: RecallSubagentResult;
-  fallbackSearchDebug?: ActiveMemorySearchDebug;
-  fallbackHasUsableMemoryResult?: boolean;
-  elapsedMs: number;
-  maxSummaryChars: number;
-}): ActiveRecallResult {
-  const { rawReply, resultStatus } = params.subagentResult;
-  const searchDebug = params.subagentResult.searchDebug ?? params.fallbackSearchDebug;
-  const summary = truncateSummary(normalizeActiveSummary(rawReply) ?? "", params.maxSummaryChars);
-  const hasUsableMemoryResult =
-    params.subagentResult.hasUsableMemoryResult === true ||
-    params.fallbackHasUsableMemoryResult === true;
-  const hasUnavailableMemorySearchResult =
-    params.subagentResult.hasUnavailableMemorySearchResult === true;
-  const canUseSummary = hasUsableMemoryResult;
-  return summary.length > 0 && canUseSummary
-    ? {
-        status: "ok",
-        elapsedMs: params.elapsedMs,
-        rawReply,
-        summary,
-        searchDebug,
-      }
-    : resultStatus === "failed"
-      ? {
-          status: "failed",
-          elapsedMs: params.elapsedMs,
-          summary: null,
-          searchDebug,
-        }
-      : resultStatus === "unavailable" ||
-          isUnavailableMemorySearchDebug(searchDebug) ||
-          hasUnavailableMemorySearchResult
-        ? {
-            status: "unavailable",
-            elapsedMs: params.elapsedMs,
-            summary: null,
-            searchDebug,
-          }
-        : {
-            status: "no_relevant_memory",
-            elapsedMs: params.elapsedMs,
-            summary: null,
-            searchDebug,
-          };
-}
-
-function resetActiveMemoryTranscriptForTests(): void {
-  timeoutPartialDataGraceMs = TIMEOUT_PARTIAL_DATA_GRACE_MS;
-}
-
-function setTimeoutPartialDataGraceMsForTests(value: number): void {
-  timeoutPartialDataGraceMs = Math.max(0, Math.floor(value));
-}
-
 export {
-  attachPartialTimeoutData,
-  buildSubagentRecallResult,
-  buildTimeoutRecallResult,
   createActiveMemoryHookDeadline,
+  extractActiveMemorySearchDebugFromSessionRecord,
+  extractToolResultNameFromSessionRecord,
   fileTranscriptSource,
+  hasTerminalUnavailableMemoryResultInSessionRecord,
+  hasUnavailableMemoryResultInSessionRecord,
   hasUsableMemoryResultInSessionRecord,
   isUnavailableMemorySearchDebug,
-  readActiveMemorySearchDebug,
-  readActiveMemorySearchDebugFromRunResult,
-  readActiveMemorySessionFileFromRunResult,
-  readMergedActiveMemoryTranscriptState,
-  readMemoryToolResultEvidence,
-  readPartialAssistantText,
-  readPartialAssistantTextFromSources,
-  readPartialTimeoutData,
-  resetActiveMemoryTranscriptForTests,
-  setTimeoutPartialDataGraceMsForTests,
+  resolveTranscriptReadLimits,
+  streamActiveMemoryTranscriptRecords,
   transcriptSourceFromReturnedSessionFile,
-  watchTerminalMemorySearchResult,
 };
